+++ src/Components/NewsLetter/NewsLetter.js
@@ -125,10 +125,7 @@ class DefaultNewsLetter extends Component {
   };
 
   postSubmit = () => {
-    window.Pelcro.paywall.decrementPageViewFrequency(
-      this.product?.paywall?.newsletter_extra_visits,
-      this.product?.paywall?.frequency_limit
-    );
+    window.Pelcro.paywall.setNewsletterExtraVisits(true);
     this.props.setView("meter");
   };