+++ src/services/Subscription/Payment.service.js
@@ -19,7 +19,7 @@ export const PAYMENT_TYPES = {
 
 export class Payment {
   /**
-   * @param {(StripeGateway|PaypalGateway|VantivGateway)} paymentGateway
+   * @param {(StripeGateway|PaypalGateway|VantivGateway|TapGateway)} paymentGateway
    */
   constructor(paymentGateway) {
     if (this.#isPaymentGatewayInvalid(paymentGateway)) {
@@ -85,7 +85,8 @@ export class Payment {
       !(
         gateway instanceof StripeGateway ||
         gateway instanceof PaypalGateway ||
-        gateway instanceof VantivGateway
+        gateway instanceof VantivGateway ||
+        gateway instanceof TapGateway
       )
     );
   };
@@ -106,7 +107,8 @@ export class Payment {
 const PAYMENT_GATEWAYS_ENUM = {
   stripe: "stripe",
   paypal: "braintree",
-  vantiv: "vantiv"
+  vantiv: "vantiv",
+  tap: "tap"
 };
 
 /**
@@ -487,6 +489,266 @@ export class VantivGateway {
 
       default:
         console.error("Unsupported payment method: vantiv Gateway");
+    }
+  };
+
+  /**
+   * Create a new subscription
+   * @param {paymentOptions} options payment options
+   * @param {executeCallback} callback
+   * @return {void}
+   */
+  #createSubscription = (options, callback) => {
+    const {
+      token,
+      plan,
+      couponCode,
+      product,
+      quantity = 1,
+      addressId,
+      isExistingSource
+    } = options;
+
+    const params = isExistingSource
+      ? {
+          source_id: token
+        }
+      : {
+          payment_gateway: this.#paymentGateway,
+          gateway_token: token
+        };
+
+    window.Pelcro.subscription.create(
+      {
+        quantity,
+        auth_token: window.Pelcro.user.read().auth_token,
+        plan_id: plan.id,
+        campaign_key:
+          window.Pelcro.helpers.getURLParameter("campaign_key"),
+        coupon_code: couponCode,
+        address_id: product.address_required ? addressId : null,
+        ...params
+      },
+      (err, res) => {
+        callback(err, res);
+      }
+    );
+  };
+
+  /**
+   * Renews a subscription
+   * @param {paymentOptions} options payment options
+   * @param {executeCallback} callback
+   * @return {void}
+   */
+  #renewSubscription = (options, callback) => {
+    const {
+      subscriptionIdToRenew,
+      token,
+      plan,
+      couponCode,
+      product,
+      addressId,
+      isExistingSource
+    } = options;
+
+    const params = isExistingSource
+      ? {
+          source_id: token
+        }
+      : {
+          payment_gateway: this.#paymentGateway,
+          gateway_token: token
+        };
+
+    window.Pelcro.subscription.renew(
+      {
+        auth_token: window.Pelcro.user.read().auth_token,
+        plan_id: plan.id,
+        coupon_code: couponCode,
+        campaign_key:
+          window.Pelcro.helpers.getURLParameter("campaign_key"),
+        subscription_id: subscriptionIdToRenew,
+        address_id: product.address_required ? addressId : null,
+        ...params
+      },
+      (err, res) => {
+        callback(err, res);
+      }
+    );
+  };
+
+  /**
+   * Create a new gifted subscription
+   * @param {paymentOptions} options payment options
+   * @param {executeCallback} callback
+   * @return {void}
+   */
+  #createGiftedSubscription = (options, callback) => {
+    const {
+      token,
+      plan,
+      couponCode,
+      product,
+      giftRecipient,
+      quantity = 1,
+      addressId,
+      isExistingSource
+    } = options;
+
+    const params = isExistingSource
+      ? {
+          source_id: token
+        }
+      : {
+          payment_gateway: this.#paymentGateway,
+          gateway_token: token
+        };
+
+    window.Pelcro.subscription.create(
+      {
+        quantity,
+        auth_token: window.Pelcro.user.read().auth_token,
+        plan_id: plan.id,
+        coupon_code: couponCode,
+        campaign_key:
+          window.Pelcro.helpers.getURLParameter("campaign_key"),
+        gift_recipient_email: giftRecipient.email,
+        gift_recipient_first_name: giftRecipient?.firstName,
+        gift_recipient_last_name: giftRecipient?.lastName,
+        gift_start_date: giftRecipient?.startDate,
+        gift_message: giftRecipient?.giftMessage,
+        address_id: product.address_required ? addressId : null,
+        ...params
+      },
+      (err, res) => {
+        callback(err, res);
+      }
+    );
+  };
+
+  /**
+   * Renews a gifted subscription
+   * @param {paymentOptions} options payment options
+   * @param {executeCallback} callback
+   * @return {void}
+   */
+  #renewGiftedSubscription = (options, callback) => {
+    const {
+      subscriptionIdToRenew,
+      token,
+      product,
+      plan,
+      couponCode,
+      addressId,
+      isExistingSource
+    } = options;
+
+    const params = isExistingSource
+      ? {
+          source_id: token
+        }
+      : {
+          payment_gateway: this.#paymentGateway,
+          gateway_token: token
+        };
+
+    window.Pelcro.subscription.renewGift(
+      {
+        auth_token: window.Pelcro.user.read().auth_token,
+        plan_id: plan.id,
+        coupon_code: couponCode,
+        subscription_id: subscriptionIdToRenew,
+        address_id: product.address_required ? addressId : null,
+        ...params
+      },
+      (err, res) => {
+        callback(err, res);
+      }
+    );
+  };
+
+  /**
+   * purchase an Ecommerce order
+   * @param {paymentOptions} options payment options
+   * @param {executeCallback} callback
+   * @return {void}
+   */
+  #purchaseEcommerceOrder = (options, callback) => {
+    const { token, items, couponCode, addressId, isExistingSource } =
+      options;
+
+    const params = isExistingSource
+      ? {
+          source_id: token
+        }
+      : {
+          payment_gateway: this.#paymentGateway,
+          gateway_token: token
+        };
+
+    window.Pelcro.ecommerce.order.create(
+      {
+        items,
+        coupon_code: couponCode,
+        campaign_key:
+          window.Pelcro.helpers.getURLParameter("campaign_key"),
+        ...params,
+        ...(addressId && { address_id: addressId })
+      },
+      (err, res) => {
+        callback(err, res);
+      }
+    );
+  };
+
+  #payInvoice = (options, callback) => {
+    const { token, invoiceId } = options;
+
+    const params = options.isExistingSource
+      ? {
+          source_id: token,
+          invoice_id: invoiceId
+        }
+      : {
+          payment_gateway: this.#paymentGateway,
+          gateway_token: token,
+          invoice_id: invoiceId
+        };
+
+    window.Pelcro.invoice.pay(params, (err, res) => {
+      callback(err, res);
+    });
+  };
+}
+export class TapGateway {
+  #paymentGateway = PAYMENT_GATEWAYS_ENUM["tap"];
+
+  /**
+   * Payment execution method
+   * @param {paymentOptions} options payment options
+   * @param {executeCallback} callback
+   * @return {void}
+   */
+  execute = (options, callback) => {
+    const types = PAYMENT_TYPES;
+
+    switch (options.type) {
+      case types.CREATE_SUBSCRIPTION:
+        return this.#createSubscription(options, callback);
+      case types.RENEW_SUBSCRIPTION:
+        return this.#renewSubscription(options, callback);
+      case types.CREATE_GIFTED_SUBSCRIPTION:
+        return this.#createGiftedSubscription(options, callback);
+      case types.RENEW_GIFTED_SUBSCRIPTION:
+        return this.#renewGiftedSubscription(options, callback);
+      case types.PURCHASE_ECOMMERCE_ORDER:
+        return this.#purchaseEcommerceOrder(options, callback);
+      case types.PAY_INVOICE:
+        return this.#payInvoice(options, callback);
+
+      default:
+        console.error("Unsupported payment method: tap Gateway");
     }
   };